AFDP-2915 Develop Functionality for Correspondence Admin

See merge request !989

Former-commit-id: ed24c6d5d454d305dc6c931f5b41e5e452286dc0

+++ acm-plugins/acm-default-plugins/acm-admin-plugin/src/main/java/com/armedia/acm/plugins/admin/web/api/CorrespondenceTemplateAPIController.java
@@ -4,8 +4,10 @@ import com.armedia.acm.correspondence.model.CorrespondenceTemplate;
 import com.armedia.acm.correspondence.service.CorrespondenceService;
 import com.armedia.acm.plugins.admin.model.CorrespondenceTemplateRequestResponse;
 
+import org.apache.commons.io.FileUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,8 +17,12 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import java.io.File;
 import java.io.IOException;
+import java.util.Date;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * @author Lazo Lazarev a.k.a. Lazarius Borg @ zerogravity Jan 27, 2017
@@ -28,6 +34,14 @@ public class CorrespondenceTemplateAPIController
 {
 
     private CorrespondenceService correspondenceService;
+
+    @RequestMapping(value = "/templates", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    @ResponseBody
+    public List<CorrespondenceTemplateRequestResponse> getAllTemplates()
+    {
+        return correspondenceService.getAllTemplates().stream().map(t -> mapTemplateToResponse(Optional.of(t)))
+                .collect(Collectors.toList());
+    }
 
     @RequestMapping(value = "/template/{templateFileName:.+}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
@@ -41,15 +55,23 @@ public class CorrespondenceTemplateAPIController
     public CorrespondenceTemplateRequestResponse deleteTemplate(@PathVariable(value = "templateFileName") String templateFileName)
             throws IOException
     {
-        return mapTemplateToResponse(correspondenceService.deleteTemplate(templateFileName));
+        File templatesDir = new File(System.getProperty("user.home") + "/.arkcase/acm/correspondenceTemplates");
+        File template = new File(templatesDir, templateFileName);
+        if (FileUtils.deleteQuietly(template))
+        {
+            return mapTemplateToResponse(correspondenceService.deleteTemplate(templateFileName));
+        } else
+        {
+            throw new CorrespondenceTemplateNotFoundException();
+        }
     }
 
     @RequestMapping(value = "/template", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public CorrespondenceTemplateRequestResponse updateTemplate(@RequestBody CorrespondenceTemplateRequestResponse request)
-            throws IOException
+    public CorrespondenceTemplateRequestResponse updateTemplate(@RequestBody CorrespondenceTemplateRequestResponse request,
+            Authentication auth) throws IOException
     {
-        return mapTemplateToResponse(correspondenceService.updateTemplate(mapRequestToTemplate(request)));
+        return mapTemplateToResponse(correspondenceService.updateTemplate(mapRequestToTemplate(request, auth)));
     }
 
     @ExceptionHandler(CorrespondenceTemplateNotFoundException.class)
@@ -64,10 +86,11 @@ public class CorrespondenceTemplateAPIController
      */
     private CorrespondenceTemplateRequestResponse mapTemplateToResponse(Optional<CorrespondenceTemplate> templateHolder)
     {
-        CorrespondenceTemplate template = templateHolder.orElse(new CorrespondenceTemplate());
+        CorrespondenceTemplate template = templateHolder.orElseThrow(CorrespondenceTemplateNotFoundException::new);
 
         CorrespondenceTemplateRequestResponse response = new CorrespondenceTemplateRequestResponse();
 
+        response.setDisplayName(template.getDisplayName());
         response.setDocumentType(template.getDocumentType());
         response.setTemplateFilename(template.getTemplateFilename());
         if (template.getQuery() != null)
@@ -78,6 +101,9 @@ public class CorrespondenceTemplateAPIController
         response.setTemplateSubstitutionVariables(template.getTemplateSubstitutionVariables());
         response.setDateFormatString(template.getDateFormatString());
         response.setNumberFormatString(template.getNumberFormatString());
+        response.setActivated(template.isActivated());
+        response.setModifier(template.getModifier());
+        response.setModified(template.getModified());
 
         return response;
     }
@@ -86,10 +112,11 @@ public class CorrespondenceTemplateAPIController
      * @param request
      * @return
      */
-    private CorrespondenceTemplate mapRequestToTemplate(CorrespondenceTemplateRequestResponse request)
+    private CorrespondenceTemplate mapRequestToTemplate(CorrespondenceTemplateRequestResponse request, Authentication auth)
     {
         CorrespondenceTemplate template = new CorrespondenceTemplate();
 
+        template.setDisplayName(request.getDisplayName());
         template.setDocumentType(request.getDocumentType());
         template.setTemplateFilename(request.getTemplateFilename());
         template.setQuery(correspondenceService.getQueryByBeanId(request.getCorrespondenceQueryBeanId())
@@ -97,6 +124,9 @@ public class CorrespondenceTemplateAPIController
         template.setTemplateSubstitutionVariables(request.getTemplateSubstitutionVariables());
         template.setDateFormatString(request.getDateFormatString());
         template.setNumberFormatString(request.getNumberFormatString());
+        template.setActivated(request.isActivated());
+        template.setModifier(auth.getName());
+        template.setModified(new Date());
 
         return template;
     }
